[8] Fix remaining CodeFactor style issues
https://github.com/Dangerdan9631/DangrLib/issues/8

+++ src/Dangr.Math/Math/MathHelper.cs
@@ -16,52 +16,52 @@ namespace Dangr.Math
         /// <summary>
         ///     Defines the value of E.
         /// </summary>
-        public const float E = 2.71828182845904523536f;
+        public const double E = 2.71828182845904523536;
 
         /// <summary>
         ///     Defines the base-10 logarithm of E.
         /// </summary>
-        public const float Log10E = 0.434294482f;
+        public const double Log10E = 0.434294482;
 
         /// <summary>
         ///     Defines the base-2 logarithm of E.
         /// </summary>
-        public const float Log2E = 1.442695041f;
+        public const double Log2E = 1.442695041;
 
         /// <summary>
         ///     Defines the value of Pi.
         /// </summary>
-        public const float Pi = 3.141592653589793238462643383279502884197169399f;
+        public const double Pi = 3.141592653589793238462643383279502884197169399;
 
         /// <summary>
         ///     Defines the value of Pi divided by two.
         /// </summary>
-        public const float PiOver2 = MathHelper.Pi/2;
+        public const double PiOver2 = MathHelper.Pi / 2.0;
 
         /// <summary>
         ///     Defines the value of Pi divided by three.
         /// </summary>
-        public const float PiOver3 = MathHelper.Pi/3;
+        public const double PiOver3 = MathHelper.Pi / 3.0;
 
         /// <summary>
         ///     Definesthe value of  Pi divided by four.
         /// </summary>
-        public const float PiOver4 = MathHelper.Pi/4;
+        public const double PiOver4 = MathHelper.Pi / 4.0;
 
         /// <summary>
         ///     Defines the value of Pi divided by six.
         /// </summary>
-        public const float PiOver6 = MathHelper.Pi/6;
+        public const double PiOver6 = MathHelper.Pi / 6.0;
 
         /// <summary>
         ///     Defines the value of Pi multiplied by two.
         /// </summary>
-        public const float TwoPi = 2*MathHelper.Pi;
+        public const double TwoPi = 2.0 * MathHelper.Pi;
 
         /// <summary>
         ///     Defines the value of Pi multiplied by 3 and divided by two.
         /// </summary>
-        public const float ThreePiOver2 = 3*MathHelper.Pi/2;
+        public const double ThreePiOver2 = 3.0 * MathHelper.Pi / 2.0;
 
         /// <summary>
         ///     Clamps a number between a minimum and a maximum.
